+++ src/senaite/core/catalog/base_catalog.py
@@ -57,7 +57,7 @@ INDEXES = [
     ("path", "", "ExtendedPathIndex"),
     ("portal_type", "", "FieldIndex"),
     ("review_state", "", "FieldIndex"),
-    ("title", "Title", "FieldIndex"),
+    ("title", "", "FieldIndex"),
     ("UID", "", "UUIDIndex"),
 ]